feat: страница вопросов с модульными стилями
