+++ src/components/Questions/Questions.js
@@ -1,19 +1,25 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import MainTitle from '../MainTitle/MainTitle';
-import TagList from './TagList/TagList';
+import Tag from './Tag/Tag';
 import Question from './Question/Question';
 import QuestionForm from './QuestionForm/QuestionForm';
 import Preloader from '../Preloader/Preloader';
 
 import api from '../../utils/Api';
 
-import './Questions.scss';
+import styles from './Questions.module.scss';
 
 export default function Questions({ loggedIn, activeRubrics, selectRubric }) {
   const [isLoading, setIsLoading] = React.useState(true);
   const [tags, setTags] = React.useState([]);
   const [questions, setQuestions] = React.useState([]);
+
+  const [activeFilter, setActiveFilter] = React.useState('');
+
+  function onFilterChange(tag) {
+    setActiveFilter(tag);
+  }
 
   React.useEffect(() => {
     Promise.all([
@@ -36,12 +42,27 @@ export default function Questions({ loggedIn, activeRubrics, selectRubric }) {
   }, []);
 
   return isLoading ? (<Preloader />) : (
-    <main className="main">
-      <section className="lead page__section">
+    <main className={styles.main}>
+      <section className={`${styles.lead} ${styles.page__section}`}>
         <MainTitle title="Ответы на вопросы" />
-        <TagList tags={tags} selectRubric={selectRubric} />
+        <div className={`${styles.tags} ${styles['tags_content_long-list']}`}>
+          <ul className={`${styles.tags__list} ${styles.tags__list_type_long}`}>
+            {
+              tags.map((tag, i) => (
+                <Tag
+                  key={tag.id}
+                  isActive={i === 0}
+                  name={tag.name}
+                  selectRubric={selectRubric}
+                  onFilterChange={onFilterChange}
+                  activeFilter={activeFilter}
+                />
+              ))
+            }
+          </ul>
+        </div>
       </section>
-      <section className="questions page__section">
+      <section className={`${styles.questions} ${styles.page__section}`}>
         {
           questions.map((question) => (
             <Question
