--- src/components/Questions/Questions.scss
@@ -1,18 +1,0 @@
-@import '../../scss/mixins.scss';
-
-.questions {
-  max-width: 970px;
-  margin: 0 auto;
-}
-
-@include atDesktop1024 {
-  .questions {
-    padding-top: 50px;
-  }
-}
-
-@include atMobile {
-  .questions {
-    padding-top: 30px;
-  }
-}
