+++ src/components/Questions/Tag/Tag.module.scss
@@ -1,5 +1,5 @@
-@import '../../../../scss/variables.scss';
-@import '../../../../scss/mixins.scss';
+@import '../../../scss/variables.scss';
+@import '../../../scss/mixins.scss';
 
 .button {
   height: 48px;
